Combine into single "AbstractDartsModel"
