+++ src/main/kotlin/dartzee/ai/AbstractDartsModel.kt
@@ -5,16 +5,16 @@ import dartzee.`object`.SegmentType
 import dartzee.`object`.getSegmentTypeForClockType
 import dartzee.core.obj.HashMapCount
 import dartzee.core.util.*
+import dartzee.logging.CODE_AI_ERROR
 import dartzee.logging.CODE_SIMULATION_FINISHED
 import dartzee.logging.CODE_SIMULATION_STARTED
 import dartzee.logging.LoggingCode
 import dartzee.screen.Dartboard
 import dartzee.screen.dartzee.DartzeeDartboard
 import dartzee.screen.game.dartzee.SegmentStatus
-import dartzee.utils.DartsDatabaseUtil
-import dartzee.utils.InjectedThings
+import dartzee.utils.*
 import dartzee.utils.InjectedThings.logger
-import dartzee.utils.getAverage
+import org.apache.commons.math3.distribution.NormalDistribution
 import org.w3c.dom.Element
 import java.awt.Point
 import java.util.*
@@ -24,11 +24,20 @@ enum class DartzeePlayStyle {
     AGGRESSIVE
 }
 
-abstract class AbstractDartsModel
+class AbstractDartsModel
 {
-    var scoringDart = 20
+    //Modelling
+    private val mean = 0
+
+    var standardDeviation = 50.0
+    var standardDeviationDoubles = -1.0
+    var standardDeviationCentral = -1.0
+
+    var distribution: NormalDistribution? = null
+    var distributionDoubles: NormalDistribution? = null
 
     //X01
+    var scoringDart = 20
     var hmScoreToDart = mutableMapOf<Int, Dart>()
     var mercyThreshold = -1
 
@@ -38,16 +47,6 @@ abstract class AbstractDartsModel
 
     //Dartzee
     var dartzeePlayStyle = DartzeePlayStyle.CAUTIOUS
-
-    /**
-     * Abstract methods
-     */
-    abstract fun getModelName(): String
-    abstract fun getType(): Int
-    abstract fun writeXmlSpecific(rootElement: Element)
-    abstract fun readXmlSpecific(root: Element)
-    abstract fun throwDartAtPoint(pt: Point, dartboard: Dartboard): Point
-    abstract fun getProbabilityWithinRadius(radius: Double): Double
 
     /**
      * Non-abstract stuff
@@ -138,7 +137,7 @@ abstract class AbstractDartsModel
     fun writeXml(): String
     {
         val xmlDoc = XmlUtil.factoryNewDocument()
-        val rootElement = xmlDoc.createRootElement(getModelName())
+        val rootElement = xmlDoc.createRootElement("Gaussian")
 
         if (scoringDart != 20)
         {
@@ -345,11 +344,119 @@ abstract class AbstractDartsModel
         hmDartNoToStopThreshold[dartNo] = stopThreshold
     }
 
+    fun throwDartAtPoint(pt: Point, dartboard: Dartboard): Point
+    {
+        if (standardDeviation == 0.0)
+        {
+            logger.error(CODE_AI_ERROR, "Gaussian model with SD of 0 - this shouldn't be possible!")
+            return pt
+        }
+
+        val (radius, angle) = calculateRadiusAndAngle(pt, dartboard)
+
+        return translatePoint(pt, radius, angle)
+    }
+
+    data class DistributionSample(val radius: Double, val theta: Double)
+    fun calculateRadiusAndAngle(pt: Point, dartboard: Dartboard): DistributionSample
+    {
+        //Averaging logic
+        val radius = sampleRadius(pt, dartboard)
+
+        //Generate the angle
+        val theta = generateAngle(pt, dartboard)
+        val sanitisedAngle = sanitiseAngle(theta)
+
+        return DistributionSample(radius, sanitisedAngle)
+    }
+    private fun sampleRadius(pt: Point, dartboard: Dartboard): Double
+    {
+        val distribution = getDistributionToUse(pt, dartboard)!!
+        return distribution.sample()
+    }
+    private fun sanitiseAngle(angle: Double): Double
+    {
+        return when
+        {
+            angle < 0 -> angle + 360
+            angle > 360 -> angle - 360
+            else -> angle
+        }
+    }
+
+    private fun getDistributionToUse(pt: Point, dartboard: Dartboard): NormalDistribution?
+    {
+        return if (dartboard.isDouble(pt) && distributionDoubles != null) distributionDoubles else distribution
+    }
+
+    private fun generateAngle(pt: Point, dartboard: Dartboard): Double
+    {
+        if (dartboard.isDouble(pt) || standardDeviationCentral == 0.0)
+        {
+            //Just pluck a number from 0-360.
+            return generateRandomAngle()
+        }
+
+        //Otherwise, we have a Normal Distribution to use to generate an angle more likely to be into the dartboard (rather than out of it)
+        val angleToAvoid = getAngleForPoint(pt, dartboard.centerPoint)
+        val angleTowardsCenter = (angleToAvoid + 180) % 360
+        val angleDistribution = NormalDistribution(angleTowardsCenter, standardDeviationCentral)
+        return angleDistribution.sample()
+    }
+
+    fun writeXmlSpecific(rootElement: Element)
+    {
+        rootElement.setAttribute(ATTRIBUTE_STANDARD_DEVIATION, "" + standardDeviation)
+
+        if (standardDeviationDoubles > 0)
+        {
+            rootElement.setAttribute(ATTRIBUTE_STANDARD_DEVIATION_DOUBLES, "" + standardDeviationDoubles)
+        }
+
+        if (standardDeviationCentral > 0)
+        {
+            rootElement.setAttribute(ATTRIBUTE_STANDARD_DEVIATION_CENTRAL, "" + standardDeviationCentral)
+        }
+    }
+
+    fun readXmlSpecific(root: Element)
+    {
+        val sd = root.getAttributeDouble(ATTRIBUTE_STANDARD_DEVIATION)
+        val sdDoubles = root.getAttributeDouble(ATTRIBUTE_STANDARD_DEVIATION_DOUBLES)
+        val sdCentral = root.getAttributeDouble(ATTRIBUTE_STANDARD_DEVIATION_CENTRAL)
+
+        populate(sd, sdDoubles, sdCentral)
+    }
+
+    fun getProbabilityWithinRadius(radius: Double): Double
+    {
+        return distribution!!.probability(-radius, radius)
+    }
+
+    fun populate(standardDeviation: Double, standardDeviationDoubles: Double, standardDeviationCentral: Double)
+    {
+        this.standardDeviation = standardDeviation
+        this.standardDeviationDoubles = standardDeviationDoubles
+        this.standardDeviationCentral = standardDeviationCentral
+
+        distribution = NormalDistribution(mean.toDouble(), standardDeviation)
+        if (standardDeviationDoubles > 0)
+        {
+            distributionDoubles = NormalDistribution(mean.toDouble(), standardDeviationDoubles)
+        }
+        else
+        {
+            distributionDoubles = null
+        }
+    }
+
     companion object
     {
-        const val TYPE_NORMAL_DISTRIBUTION = 2
-
         const val DARTS_MODEL_NORMAL_DISTRIBUTION = "Simple Gaussian"
+
+        const val ATTRIBUTE_STANDARD_DEVIATION = "StandardDeviation"
+        const val ATTRIBUTE_STANDARD_DEVIATION_DOUBLES = "StandardDeviationDoubles"
+        const val ATTRIBUTE_STANDARD_DEVIATION_CENTRAL = "StandardDeviationCentral"
 
         const val TAG_SETUP_DART = "SetupDart"
         const val TAG_GOLF_AIM = "GolfAim"
@@ -363,35 +470,6 @@ abstract class AbstractDartsModel
 
         private const val SCORING_DARTS_TO_THROW = 20000
         private const val DOUBLE_DARTS_TO_THROW = 20000
-
-        /**
-         * Static methods
-         */
-        fun factoryForType(type: Int): AbstractDartsModel?
-        {
-            return when (type)
-            {
-                TYPE_NORMAL_DISTRIBUTION -> DartsModelNormalDistribution()
-                else -> null
-            }
-        }
-
-        fun getModelDescriptions(): Vector<String>
-        {
-            val models = Vector<String>()
-            models.add(DARTS_MODEL_NORMAL_DISTRIBUTION)
-            return models
-        }
-
-        fun getStrategyDesc(type: Int): String?
-        {
-            return when(type)
-            {
-                TYPE_NORMAL_DISTRIBUTION -> DARTS_MODEL_NORMAL_DISTRIBUTION
-                else -> null
-            }
-        }
-
 
         /**
          * Get the application-wide default thing to aim for, which applies to any score of 60 or less
