+++ src/test/kotlin/dartzee/screen/ai/TestAIConfigurationGolfDartPanel.kt
@@ -4,7 +4,7 @@ import com.github.alexburlton.swingtest.getChild
 import com.github.alexburlton.swingtest.shouldBeDisabled
 import com.github.alexburlton.swingtest.shouldBeEnabled
 import dartzee.`object`.SegmentType
-import dartzee.ai.DartsModelNormalDistribution
+import dartzee.ai.AbstractDartsModel
 import dartzee.core.bean.ComboBoxItem
 import dartzee.core.bean.items
 import dartzee.core.bean.selectedItemTyped
@@ -41,7 +41,7 @@ class TestAIConfigurationGolfDartPanel: AbstractTest()
     @Test
     fun `Should initialise from a model correctly`()
     {
-        val model = DartsModelNormalDistribution()
+        val model = AbstractDartsModel()
         model.hmDartNoToSegmentType[1] = SegmentType.TREBLE
         model.hmDartNoToSegmentType[2] = SegmentType.INNER_SINGLE
 
@@ -62,7 +62,7 @@ class TestAIConfigurationGolfDartPanel: AbstractTest()
     @Test
     fun `Should populate model correctly`()
     {
-        val model = DartsModelNormalDistribution()
+        val model = AbstractDartsModel()
 
         val panel = AIConfigurationGolfDartPanel(1)
         panel.getChild<JSpinner>().value = 2
